Call tree index correction
It turns out that the format is not saved in either depth-first or
breadth-first format, but some sort of hybrid.  It stores a node, then
its children, then recurses on each child.

I have no idea what this is called, but it has a basic recursive
structure.  Anyway, it seems to work now.