Reading over locations (not locationgroups)
Previously, the number of bytes read was based on the number of location
groups, rather than the number of locations.  This worked for
single-threaded programs, but failed for multithreaded ones (e.g.
OpenMP).

We now correctly count the number of locations in each location group
and read the correct number of bytes.