index support; exclusive indexing
The #.index file for each metric is now used to populate the metric data
across the calltree nodes.

Previously, we assumed that all values existed.  However, only values
explicitly set in the #.index file are saved in the CUBE file.  We now
use this file to parse the #.data file and populate entries.

We also support the "EXTERNAL" indexing of metrics.  This index
populates entries by sequentially assigning indices to a node, followed
by each of its children, and then recursively repeating this on each
child.  This is neither breadth-first nor depth-first, but a "fanning"
from (say) left to right.

This patch greatly improves support for Scalasca CUBE files, although
it is still not complete.