Inclusive index bugfix
The last change fixed exclusive indexing, but broke inclusive indexing,
which was previously default.  It was incorrectly replaced with the
calltree index order, which is neither exclusive nor inclusive.

This is more a quick patch, its probably better to replace
`{in|ex}clusive_index` with lists of cnodes rather than index maps.
This is just an emergency fix.