System saved as a hierarchical tree
The system info is more hierarchichal, with nodes for machines and
switches, rather than just a flat structure which the code was assuming.

This change will load in the structure as a tree, but basically breaks
everything else because now locationgrp is unpopulated.  Will address
this in a later patch.