System node as tree; print_weights bugfix
This patch replaces the assumed-flat system node layout with a
hierarchical tree.  Previously (on NCI) it was assumed that there was a
single system node, which then contained a flat list of location groups,
each containing a flat list of locations.

On more complex machines for which Score-P is more attuned, there is a
deeper hierarchy of system nodes, which can contain information about
intermediate points such as network switches.  This was causing cubex to
fail, since there were no location groups inside the top-level system
group.

We have resolved this by replacing the system node parsing with a
recursive parser.  We still force all of the location groups into a list
whose indices are not matched to their Ids.  Although at first glance
they appear to be correct, this should be addressed in the future.

----

We have also fixed a bug in the `print_weights` function, which was not
including the final rank in its averaged metric reports.  In addition to
being incorrect, this was causing runtime-fails in single-rank tests.